[charts] Keep unused plugins out of per-chart bundles
Specific charts (Line, Bar, Pie, Scatter, Radar) pass their own plugin list,
but they routed through helpers that referenced `DEFAULT_PLUGINS` as a fallback
(`plugins ?? DEFAULT_PLUGINS`, `plugins = DEFAULT_PLUGINS`). That static
reference forced the full plugin set — including `useChartClosestPoint` and
`useProgressiveRendering` — into every chart's bundle even though they never ran.

- Split the data provider into `ChartDataProviderInternal` (requires `plugins`,
  rendered by chart components) and the public `ChartsDataProvider` wrapper
  (supplies `DEFAULT_PLUGINS` for standalone/composition usage). Drop the
  fallback from `useChartsContainerProps`/`useChartsDataProviderProps`.
- Move the scatter-only Flatbush spatial-index selector out of the shared
  cartesian-axis selectors into the `useChartClosestPoint` plugin, so
  `Flatbush`/`flatqueue` no longer ship in non-scatter cartesian charts.

Public API is unchanged. Measured on the `LineChart` entrypoint:
87.6 KB → 82.9 KB gzip (-4.7 KB). Similar for Bar/Pie; Scatter keeps
closest-point and Flatbush, as it uses them.
