+++ packages/x-charts/src/ChartsDataProvider/index.ts
@@ -1,1 +1,2 @@
 export * from './ChartsDataProvider';
+export { ChartDataProviderInternal } from './ChartDataProviderInternal';
